Comprehension 2: Standardize code for problem 8
Originally I wanted to have different input values that used different
"database states", but after thinking about it that doesn't seem worth
the extra complication in the code.

I've simplified back to the style that is used for other problems, with
the main data structure being provided as a constant in the code
snippet. The input values section only has calls to the method from the
code snippet, as with previous problems.